mainServer/requirements.txt
asgiref==3.2.5
astroid==2.3.3
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
colorama==0.4.3
cryptography==38.0.3
defusedxml==0.7.1
Django==3.0.4
django-crispy-forms==1.9.0
django-froala-editor==3.1.0
djangorestframework==3.13.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.2.2
Pillow==9.2.0
Pillow-PIL==0.1.dev0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2019.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.14.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
tzdata==2022.1
urllib3==1.25.8
wrapt==1.11.2
